@@ -1,30 +1,26 @@
 {
-  "name": "toggle-tech-lab",
+  "name": "tailwind",
   "version": "0.1.0",
   "private": true,
-  "engine": { "node": "16.15.1", "npm": "8.11.0" },
   "dependencies": {
-    "@craco/craco": "^6.1.0",
-    "@tailwindcss/postcss7-compat": "^2.0.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@thelevicole/youtube-to-html5-loader": "^5.0.0",
-    "autoprefixer": "^9.8.6",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-icons": "^4.2.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.9.0",
-    "react-scripts": "4.0.1",
+    "react-scripts": "5.0.1",
+    "react-to-print": "^2.14.7",
+    "web-vitals": "^2.1.4",
     "react-twitter-embed": "^4.0.4",
     "swiper": "^9.2.0",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.2",
-    "web-vitals": "^0.2.4"
+    "react-icons": "^4.2.0",
+    "@thelevicole/youtube-to-html5-loader": "^5.0.0"
   },
   "scripts": {
-    "start": "serve -s build",
-    "build": "craco --openssl-legacy-provider build",
-    "test": "craco test",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -46,6 +42,6 @@
     ]
   },
   "devDependencies": {
-    "postcss": "^8.4.21"
+    "tailwindcss": "^3.1.3"
   }
 }
